[Clean] Fix review comments on view decomposition into several smaller sub views
